+++ Non_Graphique/BatailleNaval/nbproject/project.properties
@@ -54,7 +54,7 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-main.class=bataillenaval.BatailleNaval
+main.class=bataillenaval.controller.BatailleNaval
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
