+++ Non_Graphique/BatailleNaval/src/bataillenaval/controller/Sauvegarde.java
@@ -1,4 +1,4 @@
-package bataillenaval;
+package bataillenaval.controller;
 
 
 public class Sauvegarde {
